requirements.txt
gunicorn
uvloop
motor
requests
appdirs==1.4.4
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
h11==0.12.0
h2==3.2.0
hpack==3.0.0
httpcore==0.12.3
httptools==0.1.1
httpx==0.16.1
hyperframe==5.2.0
idna==2.10
Jinja2==2.11.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
pathspec==0.8.1
pydantic==1.7.3
python-dotenv==0.15.0
PyYAML==5.4.1
regex==2020.11.13
rfc3986==1.4.0
sniffio==1.2.0
starlette==0.13.6
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.13.3
watchgod==0.6
websockets==8.1
aiofiles
python-multipart
PyJWT
